Changed population to long. count is now per person.

+++ Pulsar4X/Pulsar4X.Tests/ProcessorTests.cs
@@ -92,7 +92,7 @@ namespace Pulsar4X.Tests
 
 
             ColonyInfoDB colonyInfo = colonyEntity.GetDataBlob<ColonyInfoDB>();
-            JDictionary<Entity, double> pop = colonyInfo.Population;
+            JDictionary<Entity, long> pop = colonyInfo.Population;
             var species = pop.Keys.ToList();
             colonyInfo.Population[species[0]] = 5; //5mil pop
